# perf: boost Speed Index and optimize bundle