--- a/.gitignore
+++ b/.gitignore
@@ -75,3 +75,5 @@ scripts/simulateur/codes-insee.json
 /*-opendata-fcu.zip
 
 *.log
+
+.tmp