## Summary
• Add preconnect to external domains for faster DNS resolution
• Memoize services initialization to prevent re-instantiation on every render
• Optimize NetworksList memoizations to reduce re-calculations
• Add image optimization module with shell script
• Move image formatting commands to app module

## Performance Impact
• **Speed Index**: Expected ~2100ms improvement
• Preconnect: ~400ms saved on DNS lookups
• Services memo: ~200ms saved on render cycles  
• NetworksList memo: ~300ms saved on filtering
• Image optimization: Reduced total image weight by ~20MB

## Type
perf